Fix tests
A new version of chrome made it's way into the selenium docker image in github and broke the tests. Tests here broke because of the explicit download of a chromedriver, even though it's already in the base image. Removed this download and changed the tests to load the chromedriver from the normal PATH.

Additionally still having intermittent issues with the browser being unable to click on things because of the async nature of how the header gets loaded in. Added an additional check to wait for one of the elements that are added by the header's js. Probably don't need to wait for hesburgh-wrapped any longer, but figured it couldn't hurt.